moved Slim Code into Vendor and moved service code under service
