+++ churchinfo/service/MailchimpService.php
@@ -4,7 +4,7 @@
  * Date: 12/9/2014
  * Time: 11:00 PM
  */
-require_once '../vendor/Mailchimp.php';
+include '../vendor/Mailchimp.php';
 
 class MailChimpService
 {
